databases/pyproject.toml
[project]
name = "databases"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.9",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.9",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "redis>=6.1.0",
    "sqlalchemy>=2.0.41",
]
